project/final/requirements.txt
flask
google-cloud-translate
google-cloud-language
gunicorn
grpcio